// www/app/api/models/AudioWaveform.ts
/* tslint:disable */
/* eslint-disable */
/**
 * FastAPI
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 0.1.0
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { exists, mapValues } from "../runtime";
/**
 *
 * @export
 * @interface AudioWaveform
 */
export interface AudioWaveform {
  /**
   *
   * @type {any}
   * @memberof AudioWaveform
   */
  data: any | null;
}

/**
 * Check if a given object implements the AudioWaveform interface.
 */
export function instanceOfAudioWaveform(value: object): boolean {
  let isInstance = true;
  isInstance = isInstance && "data" in value;

  return isInstance;
}

export function AudioWaveformFromJSON(json: any): AudioWaveform {
  return AudioWaveformFromJSONTyped(json, false);
}

export function AudioWaveformFromJSONTyped(
  json: any,
  ignoreDiscriminator: boolean,
): AudioWaveform {
  if (json === undefined || json === null) {
    return json;
  }
  return {
    data: json["data"],
  };
}

export function AudioWaveformToJSON(value?: AudioWaveform | null): any {
  if (value === undefined) {
    return undefined;
  }
  if (value === null) {
    return null;
  }
  return {
    data: value.data,
  };
}